+++ tm-tourney-manager/src/scheduling.rs
@@ -5,7 +5,7 @@ pub enum Scheduling {
     // The inherited option should register a scheduled reducer.
     // In this way it could have match groups and dependencies that it triggers multiple matches.
     // This field should be the id of the scheduled reducer.
-    Inherited(u64),
+    Inherited(u32),
 }
 
 #[cfg_attr(feature = "spacetime", derive(spacetimedb::SpacetimeType))]
